Widen the size arithmetic and drop the remaining double precision
Re-applies the yz_num_type_overflow fixes to the files the marching-observer
rewrite did not already cover. Byte counts built as frames * per_frame were int
products that promote to size_t only after wrapping: 1024 frames x 128^3 gives
a nonsense length and 4096 frames gives exactly zero, so a copy silently moves
nothing. 14 sites widened, plus the frame-indexed pointer arithmetic in
frame_interp_global.

Also makes the voxelize math explicitly single-precision and removes
gaussian_map_nd_device, the only double-returning device function, which
nothing called. Both compiled kernels now emit zero f64 instructions.

+++ src/gpuutils.cu
@@ -291,11 +291,13 @@ void aggregate_host(float *voxel_traj, float *result_grid, const int frame_numbe
                                        static_cast<size_t>(BufferSlot::TRAJ_DYNAMICS));
     tmp_grid_gpu = ctx->get_buffer_f(grid_number, static_cast<size_t>(BufferSlot::OUTPUT_GRID));
   } else {
-    CUDA_CHECK(cudaMalloc(&voxel_traj_gpu, frame_number * grid_number * sizeof(float)));
+    CUDA_CHECK(cudaMalloc(&voxel_traj_gpu,
+                          static_cast<size_t>(frame_number) * grid_number * sizeof(float)));
     CUDA_CHECK(cudaMalloc(&tmp_grid_gpu, grid_number * sizeof(float)));
   }
 
-  copy_h2d_async(ctx, voxel_traj_gpu, voxel_traj, frame_number * grid_number * sizeof(float),
+  copy_h2d_async(ctx, voxel_traj_gpu, voxel_traj,
+                 static_cast<size_t>(frame_number) * grid_number * sizeof(float),
                  BufferSlot::TRAJ_DYNAMICS, stream);
   CUDA_CHECK(cudaMemsetAsync(tmp_grid_gpu, 0, grid_number * sizeof(float), stream));
